pyproject.toml
[project]
name = "el-supervisor"
version = "0.1.0"
requires-python = ">=3.11"
description = "Proyecto de orquestación operativa y RPA para gestión académica."
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.12.1",
    "certifi==2026.1.4",
    "colorama==0.4.6",
    "distro==1.9.0",
    "et_xmlfile==2.0.0",
    "groq==1.0.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "idna==3.11",
    "markdown-it-py==4.0.0",
    "mdurl==0.1.2",
    "numpy==2.4.1",
    "pandas==3.0.0",
    "pyarrow>=19.0.0",
    "pydantic==2.12.5",
    "pydantic_core==2.41.5",
    "Pygments==2.19.2",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.2.1",
    "pywin32==311",
    "PyYAML==6.0.3",
    "rich==14.3.1",
    "six==1.17.0",
    "sniffio==1.3.1",
    "typing_extensions==4.15.0",
    "tzdata==2025.3",
    "requests==2.32.3",
    "playwright==1.50.0",
    "openpyxl==3.1.5",
]

[build-system]
requires = ["uv>=0.1.0", "setuptools>=61.0.0"]
build-backend = "setuptools.build_meta"
